usvMonitorHandlerAPI.h: Parameter bei Definition des dataRx_t Zeigers verändert, um die Einheitlich mit der dataTx_t zu halten

+++ 01 Hardware/01 Lidar/03 Software/01 source/Testen mit USVMonitor/Testen mit USVMonitor/usvMonitorHandlerAPI.h
@@ -86,7 +86,7 @@ typedef struct{
 }slaveReg_t;
 
 
-typedef uint8_t (*dataRx_t)(uint8_t* data, uint8_t length);
+typedef uint8_t (*dataRx_t)(uint8_t* data, uint16_t length);
 typedef uint8_t (*dataTx_t)(uint8_t* data, uint16_t length);
 typedef void (*wait_t)(uint32_t us);
